rename QueryBuilderOperator to QueryBuilderFilterOperator
